División de Barra de Desplazamiento Las funciones referidas a la barra principal (horizontal) y lateral (vertical) han sido traspasados a archivos propios.
